+++ pkg/kn/commands/route/list_test.go
@@ -19,6 +19,7 @@ import (
 	"testing"
 
 	"github.com/knative/client/pkg/kn/commands"
+	"github.com/knative/client/pkg/util"
 	"github.com/knative/serving/pkg/apis/serving/v1alpha1"
 	"github.com/knative/serving/pkg/apis/serving/v1beta1"
 	"gotest.tools/assert"
@@ -71,9 +72,9 @@ func TestRouteListDefaultOutput(t *testing.T) {
 	} else if !action.Matches("list", "routes") {
 		t.Errorf("Bad action %v", action)
 	}
-	assert.Check(t, commands.ContainsMultipleSubstrings(output[0], []string{"NAME", "URL", "AGE", "CONDITIONS", "TRAFFIC"}, "column header"))
-	assert.Check(t, commands.ContainsMultipleSubstrings(output[1], []string{"foo", "100% -> foo-01234"}, "value"))
-	assert.Check(t, commands.ContainsMultipleSubstrings(output[2], []string{"bar", "100% -> bar-98765"}, "value"))
+	assert.Check(t, util.ContainsAll(output[0], "NAME", "URL", "AGE", "CONDITIONS", "TRAFFIC"))
+	assert.Check(t, util.ContainsAll(output[1], "foo", "100% -> foo-01234"))
+	assert.Check(t, util.ContainsAll(output[2], "bar", "100% -> bar-98765"))
 }
 
 func TestRouteListWithTwoTargetsOutput(t *testing.T) {
@@ -88,8 +89,8 @@ func TestRouteListWithTwoTargetsOutput(t *testing.T) {
 	} else if !action.Matches("list", "routes") {
 		t.Errorf("Bad action %v", action)
 	}
-	assert.Check(t, commands.ContainsMultipleSubstrings(output[0], []string{"NAME", "URL", "AGE", "CONDITIONS", "TRAFFIC"}, "column header"))
-	assert.Check(t, commands.ContainsMultipleSubstrings(output[1], []string{"foo", "20% -> foo-01234, 80% -> foo-98765"}, "value"))
+	assert.Check(t, util.ContainsAll(output[0], "NAME", "URL", "AGE", "CONDITIONS", "TRAFFIC"))
+	assert.Check(t, util.ContainsAll(output[1], "foo", "20% -> foo-01234, 80% -> foo-98765"))
 }
 
 func createMockRouteMeta(name string) *v1alpha1.Route {
