+++ pkg/kn/commands/service/service_list_test.go
@@ -19,6 +19,7 @@ import (
 	"testing"
 
 	"github.com/knative/client/pkg/kn/commands"
+	"github.com/knative/client/pkg/util"
 	duckv1beta1 "github.com/knative/pkg/apis/duck/v1beta1"
 	v1alpha1 "github.com/knative/serving/pkg/apis/serving/v1alpha1"
 	"gotest.tools/assert"
@@ -72,9 +73,9 @@ func TestServiceListDefaultOutput(t *testing.T) {
 	} else if !action.Matches("list", "services") {
 		t.Errorf("Bad action %v", action)
 	}
-	assert.Check(t, commands.ContainsMultipleSubstrings(output[0], []string{"NAME", "DOMAIN", "GENERATION", "AGE", "CONDITIONS", "READY", "REASON"}, "column header"))
-	assert.Check(t, commands.ContainsMultipleSubstrings(output[1], []string{"foo", "foo.default.example.com", "1"}, "value"))
-	assert.Check(t, commands.ContainsMultipleSubstrings(output[2], []string{"bar", "bar.default.example.com", "2"}, "value"))
+	assert.Check(t, util.ContainsAll(output[0], "NAME", "DOMAIN", "GENERATION", "AGE", "CONDITIONS", "READY", "REASON"))
+	assert.Check(t, util.ContainsAll(output[1], "foo", "foo.default.example.com", "1"))
+	assert.Check(t, util.ContainsAll(output[2], "bar", "bar.default.example.com", "2"))
 }
 
 func createMockServiceWithParams(name, domain string, generation int64) *v1alpha1.Service {
